Rename Refund class to TransactionRefund
To keep uol standard
