+++ lib/pagseguro.rb
@@ -50,11 +50,11 @@ require "pagseguro/payment_request/response"
 require "pagseguro/payment_status"
 require "pagseguro/permission"
 require "pagseguro/request"
-require "pagseguro/refund"
-require "pagseguro/refund/request_serializer"
-require "pagseguro/refund/response"
-require "pagseguro/refund/response_serializer"
-require "pagseguro/refund/serializer"
+require "pagseguro/transaction_refund"
+require "pagseguro/transaction_refund/request_serializer"
+require "pagseguro/transaction_refund/response"
+require "pagseguro/transaction_refund/response_serializer"
+require "pagseguro/transaction_refund/serializer"
 require "pagseguro/sender"
 require "pagseguro/session"
 require "pagseguro/session/response"
