+++ lib/pagseguro/transaction_refund.rb
@@ -1,5 +1,5 @@
 module PagSeguro
-  class Refund
+  class TransactionRefund
     include Extensions::MassAssignment
 
     # Set the transaction code.
@@ -18,7 +18,7 @@ module PagSeguro
     attr_reader :errors
 
     # Calls the PagSeguro webservice and register the refund.
-    # Return PagSeguro::Refund.
+    # Return PagSeguro::TransactionRefund.
     def register
       response_request = Request.post("transactions/refunds", api_version, params)
       Response.new(response_request, self).serialize
